+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Release Notes
 
 Forthcoming
 -----------
+* [trees] setup timeout error with last behaviour name included in the error message, `#278 <https://github.com/splintered-reality/py_trees/pull/278>`_
 * [blackboard] rooted variables in namespaced clients working as designed, fixed docs, `#277 <https://github.com/splintered-reality/py_trees/pull/277>`_
 
 2.0.9 (2020-02-14)